# shopping-api/src/main/resources/shopping-openapi.yaml
openapi: 3.1.0
info:
  title: Devbooks
  description: Devbooks shopping API
  version: 1.0.0
servers:
  - url: 'http://localhost:8080'
  - url: 'https://devbooks.jsilkens.be'
tags:
  - name: Book
    description: Actions on books
  - name: Cart
    description: Actions on the shopping cart
paths:
  /book:
    get:
      tags:
        - Book
      summary: Get a paginated list of books
      operationId: getPaginatedBooks
      parameters:
        - name: page
          in: query
          required: false
          description: The page number (1-based)
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: size
          in: query
          required: false
          description: The number of items per page
          schema:
            type: integer
            default: 10
            minimum: 1
      responses:
        '200':
          description: Successfully retrieved the paginated list of books
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedBookResponseDTO'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

  /book/{isbn}:
    get:
      tags:
        - Book
      summary: Get a book by its ISBN number
      operationId: getBookByIsbn
      parameters:
        - name: isbn
          in: path
          required: true
          description: The ISBN-13 of the book to retrieve
          schema:
            type: string
            pattern: '^\d{13}$'
      responses:
        '200':
          description: Successfully retrieved the book
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDTO'
        '400':
          description: Invalid ISBN format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'

components:
  schemas:
    BookDTO:
      type: object
      required:
        - title
        - author
        - publication_year
        - isbn
        - price
      properties:
        title:
          type: string
          example: "Clean Code"
        author:
          type: string
          example: "Robert C. Martin"
        publication_year:
          type: integer
          example: 2008
        isbn:
          type: string
          example: "9780132350884"
        price:
          $ref: '#/components/schemas/Price'

    BookListItemDTO:
      type: object
      required:
        - title
        - author
        - price
        - details_link
      properties:
        title:
          type: string
          example: "Clean Code"
        author:
          type: string
          example: "Robert C. Martin"
        price:
          $ref: '#/components/schemas/Price'
        details_link:
          type: string
          example: "/api/book/9780132350884"

    PaginatedBookResponseDTO:
      type: object
      required:
        - items
        - metadata
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/BookListItemDTO'
        metadata:
          $ref: '#/components/schemas/PaginationMetadataDTO'

    PaginationMetadataDTO:
      type: object
      required:
        - current_page
        - page_size
        - total_items
        - total_pages
      properties:
        current_page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 10
        total_items:
          type: integer
          format: int64
          example: 15
        total_pages:
          type: integer
          example: 2

    Price:
      type: object
      required:
        - value
        - currency
      properties:
        value:
          type: number
          format: double
          example: 50.00
        currency:
          type: string
          example: "EUR"

    ProblemDetail:
      type: object
      properties:
        type:
          type: string
          format: uri
          description: A URI reference that identifies the problem type.
          example: "about:blank"
        title:
          type: string
          description: A short, human-readable summary of the problem type.
          example: "Not Found"
        status:
          type: integer
          description: The HTTP status code.
          example: 404
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence of the problem.
          example: "Book not found"
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence of the problem.
          example: "/api/book/9780132350884"